requirements.txt
requests>=2.31.0
python-docx>=1.1.0
anthropic>=0.25.0
google-generativeai>=0.8.0
